// src/client/signupapi.ts
export interface RequestAccessRequest {
    /**
     * 
     * @type {string}
     * @memberof RequestAccessRequest
     */
    'firstName'?: string;
    /**
     * 
     * @type {string}
     * @memberof RequestAccessRequest
     */
    'lastName'?: string;
    /**
     * 
     * @type {string}
     * @memberof RequestAccessRequest
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof RequestAccessRequest
     */
    'programmingLang'?: string;
    /**
     * 
     * @type {string}
     * @memberof RequestAccessRequest
     */
    'projectDesc'?: string;
    /**
     * 
     * @type {string}
     * @memberof RequestAccessRequest
     */
    'userName'?: string;
}